refactor(view): java support
