+++ example/view-based/src/main/java/com/neoutils/highlight/example/view/launcher/TextViewActivity.kt
@@ -1,8 +1,8 @@
-package com.neoutils.hightlight.example.view.launcher
+package com.neoutils.highlight.example.view.launcher
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
-import com.neoutils.hightlight.example.view.databinding.ActivityTextViewBinding
+import com.neoutils.highlight.example.view.databinding.ActivityTextViewBinding
 
 class TextViewActivity : AppCompatActivity() {
