+++ example/view-based/src/main/res/layout/activity_text_view.xml
@@ -6,29 +6,34 @@
         android:layout_height="match_parent"
         android:orientation="vertical"
         android:gravity="center"
-        tools:context=".launcher.TextViewActivity">
+        tools:context="com.neoutils.highlight.example.view.launcher.TextViewActivity">
 
-    <com.neoutils.hightlight.example.view.example.TextColorSchemeExample
+    <com.neoutils.highlight.example.view.example.TextColorSchemeExample
             android:textSize="18sp"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"/>
 
-    <com.neoutils.hightlight.example.view.example.BackgroundSchemeExample
+    <com.neoutils.highlight.example.view.example.BackgroundSchemeExample
             android:textSize="18sp"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"/>
 
-    <com.neoutils.hightlight.example.view.example.StyleSchemeExample
+    <com.neoutils.highlight.example.view.example.StyleSchemeExample
             android:textSize="18sp"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"/>
 
-    <com.neoutils.hightlight.example.view.example.CaptureGroupsExample
+    <com.neoutils.highlight.example.view.example.CaptureGroupsExample
             android:textSize="18sp"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"/>
 
-    <com.neoutils.hightlight.example.view.example.TextFontSchemeExample
+    <com.neoutils.highlight.example.view.example.TextFontSchemeExample
+            android:textSize="18sp"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"/>
+
+    <com.neoutils.highlight.example.view.example.JavaCaseExample
             android:textSize="18sp"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"/>
