+++ highlight/core/build.gradle.kts
@@ -7,7 +7,7 @@ group = "highlight"
 
 android {
     namespace = "com.neoutils.highlight.core"
-    compileSdk = 33
+    compileSdk = 34
 
     defaultConfig {
         minSdk = 21
