+++ highlight/core/src/main/kotlin/com/neoutils/highlight/core/utils/Match.kt
@@ -4,14 +4,18 @@ data class Match<T : Any>(
     val values: List<T?>,
 ) {
     companion object {
+
+        @JvmStatic
         fun <T : Any> fully(
             value: T
         ) = Match(listOf(value))
 
+        @JvmStatic
         fun <T : Any> groups(
             values: List<T?>
         ) = Match(values = listOf(null) + values)
 
+        @JvmStatic
         fun <T : Any> groups(
             vararg values: T?
         ) = Match(listOf(null, *values))
